Inlined most if-else branches while accounting for readability.
In some cases, inverted the conditions to short-circuit.

+++ src/main/java/meteordevelopment/meteorclient/systems/commands/commands/DropCommand.java
@@ -39,31 +39,23 @@ public class DropCommand extends Command {
 
         // Hotbar
         builder.then(literal("hotbar").executes(context -> drop(player -> {
-            for (int i = 0; i < 9; i++) {
-                InvUtils.drop().slotHotbar(i);
-            }
+            for (int i = 0; i < 9; i++) InvUtils.drop().slotHotbar(i);
         })));
 
         // Main Inv
         builder.then(literal("inventory").executes(context -> drop(player -> {
-            for (int i = 9; i < player.getInventory().main.size(); i++) {
-                InvUtils.drop().slotMain(i - 9);
-            }
+            for (int i = 9; i < player.getInventory().main.size(); i++) InvUtils.drop().slotMain(i - 9);
         })));
 
         // Hotbar and main inv
         builder.then(literal("all").executes(context -> drop(player -> {
-                    for (int i = 0; i < player.getInventory().size(); i++) {
-                        InvUtils.drop().slot(i);
-                    }
-                })));
+            for (int i = 0; i < player.getInventory().size(); i++) InvUtils.drop().slot(i);
+        })));
 
         // Armor
         builder.then(literal("armor").executes(context -> drop(player -> {
-                    for (int i = 0; i < player.getInventory().armor.size(); i++) {
-                        InvUtils.drop().slotArmor(i);
-                    }
-                })));
+            for (int i = 0; i < player.getInventory().armor.size(); i++) InvUtils.drop().slotArmor(i);
+        })));
 
         // Specific item
         builder.then(argument("item", ItemStackArgumentType.itemStack(Commands.REGISTRY_ACCESS)).executes(context -> drop(player -> {
@@ -72,9 +64,7 @@ public class DropCommand extends Command {
             if (stack == null || stack.getItem() == Items.AIR) throw NO_SUCH_ITEM.create();
 
             for (int i = 0; i < player.getInventory().size(); i++) {
-                if (stack.getItem() == player.getInventory().getStack(i).getItem()) {
-                    InvUtils.drop().slot(i);
-                }
+                if (stack.getItem() == player.getInventory().getStack(i).getItem()) InvUtils.drop().slot(i);
             }
         })));
     }
